refactor(shortcuts): restructure keyboard shortcut constants by moving to dedicated files for local and global shortcuts
