+++ packages/shared/src/index.ts
@@ -1,5 +1,5 @@
 export * from './constants.js';
-export * from './keyboard.js';
+export * from './keyboard/index.js';
 export * from './result.js';
 export * from './toasts.js';
 export * from './utils.js';
