+++ packages/shared/src/keyboard/local.ts
@@ -1,0 +1,241 @@
+/**
+ * Structured keyboard key sections for local (browser-based) shortcuts.
+ * Each section groups related keys with descriptive metadata.
+ * Keys are lowercase as returned by KeyboardEvent.key.toLowerCase()
+ */
+export const SUPPORTED_KEY_SECTIONS = [
+	{
+		title: 'Modifiers',
+		description: 'Hold with other keys',
+		keys: [
+			'control',
+			'shift',
+			'alt',
+			'meta', // Command on Mac, Windows key on PC
+			'altgraph',
+			'capslock',
+			'numlock',
+			'scrolllock',
+			'fn',
+			'fnlock',
+			'hyper',
+			'super',
+			'symbol',
+			'symbollock',
+		] as const,
+	},
+	{
+		title: 'Letters',
+		description: 'Any letter A-Z',
+		keys: [
+			'a',
+			'b',
+			'c',
+			'd',
+			'e',
+			'f',
+			'g',
+			'h',
+			'i',
+			'j',
+			'k',
+			'l',
+			'm',
+			'n',
+			'o',
+			'p',
+			'q',
+			'r',
+			's',
+			't',
+			'u',
+			'v',
+			'w',
+			'x',
+			'y',
+			'z',
+		] as const,
+	},
+	{
+		title: 'Numbers',
+		description: 'Number keys 0-9',
+		keys: ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'] as const,
+	},
+	{
+		title: 'Symbols & Punctuation',
+		description: 'Common symbols and punctuation',
+		keys: [
+			'`',
+			'~',
+			'!',
+			'@',
+			'#',
+			'$',
+			'%',
+			'^',
+			'&',
+			'*',
+			'(',
+			')',
+			'-',
+			'_',
+			'=',
+			'+',
+			'[',
+			'{',
+			']',
+			'}',
+			'\\',
+			'|',
+			';',
+			':',
+			"'",
+			'"',
+			',',
+			'<',
+			'.',
+			'>',
+			'/',
+			'?',
+		] as const,
+	},
+	{
+		title: 'Whitespace',
+		description: 'Space, enter, and tab keys',
+		keys: [' ', 'enter', 'tab'] as const,
+	},
+	{
+		title: 'Navigation',
+		description: 'Arrow and positioning keys',
+		keys: [
+			'arrowleft',
+			'arrowright',
+			'arrowup',
+			'arrowdown',
+			'home',
+			'end',
+			'pageup',
+			'pagedown',
+		] as const,
+	},
+	{
+		title: 'Editing',
+		description: 'Text editing keys',
+		keys: [
+			'backspace',
+			'delete',
+			'insert',
+			'clear',
+			'copy',
+			'cut',
+			'paste',
+			'redo',
+			'undo',
+		] as const,
+	},
+	{
+		title: 'Function Keys',
+		description: 'F1-F24 function keys',
+		keys: [
+			'f1',
+			'f2',
+			'f3',
+			'f4',
+			'f5',
+			'f6',
+			'f7',
+			'f8',
+			'f9',
+			'f10',
+			'f11',
+			'f12',
+			'f13',
+			'f14',
+			'f15',
+			'f16',
+			'f17',
+			'f18',
+			'f19',
+			'f20',
+			'f21',
+			'f22',
+			'f23',
+			'f24',
+		] as const,
+	},
+	{
+		title: 'Special Keys',
+		description: 'System and browser keys',
+		keys: [
+			'escape',
+			'contextmenu',
+			'pause',
+			'break',
+			'printscreen',
+			'help',
+			'browserback',
+			'browserforward',
+			'browserhome',
+			'browserrefresh',
+			'browsersearch',
+			'browserstop',
+			'browserfavorites',
+		] as const,
+	},
+	{
+		title: 'Media Keys',
+		description: 'Media control keys',
+		keys: [
+			'mediaplaypause',
+			'mediaplay',
+			'mediapause',
+			'mediastop',
+			'mediatracknext',
+			'mediatrackprevious',
+			'volumeup',
+			'volumedown',
+			'volumemute',
+		] as const,
+	},
+	{
+		title: 'Other Keys',
+		description: 'Additional special keys',
+		keys: [
+			'dead',
+			'unidentified',
+			'process',
+			'compose',
+			'accept',
+			'again',
+			'attn',
+			'cancel',
+			'execute',
+			'find',
+			'finish',
+			'props',
+			'select',
+			'zoomout',
+			'zoomin',
+		] as const,
+	},
+] as const;
+
+/**
+ * Flattened array of all supported local shortcut keys for compatibility.
+ * This is derived from SUPPORTED_KEY_SECTIONS for use in validation and type guards.
+ */
+export const ALL_SUPPORTED_KEYS = [
+	...SUPPORTED_KEY_SECTIONS[0].keys, // Modifiers
+	...SUPPORTED_KEY_SECTIONS[1].keys, // Letters
+	...SUPPORTED_KEY_SECTIONS[2].keys, // Numbers
+	...SUPPORTED_KEY_SECTIONS[3].keys, // Symbols
+	...SUPPORTED_KEY_SECTIONS[4].keys, // Whitespace
+	...SUPPORTED_KEY_SECTIONS[5].keys, // Navigation
+	...SUPPORTED_KEY_SECTIONS[6].keys, // Editing
+	...SUPPORTED_KEY_SECTIONS[7].keys, // Function
+	...SUPPORTED_KEY_SECTIONS[8].keys, // Special
+	...SUPPORTED_KEY_SECTIONS[9].keys, // Media
+	...SUPPORTED_KEY_SECTIONS[10].keys, // Other
+] as const satisfies (typeof SUPPORTED_KEY_SECTIONS)[number]['keys'][number][];
+
+export type SupportedKey = (typeof ALL_SUPPORTED_KEYS)[number];
